+++ rust/Cargo.toml
@@ -52,6 +52,7 @@ version = "0.1.0"
 Inflector = "0.11.4"
 anyhow = "1.0"
 async-trait = "0.1"
+async-rwlock = "1.3"
 auto_impl = "1.0"
 backtrace = "0.3"
 base64 = "0.21.2"
