+++ rust/hyperlane-core/src/rpc_clients/mod.rs
@@ -1,0 +1,4 @@
+pub use self::{error::*, fallback::*};
+
+mod error;
+mod fallback;
